+++ SyncPod/coordinator/BaseCoordinator.swift
@@ -1,0 +1,11 @@
+//
+//  BaseCoordinator.swift
+//  SyncPod
+//
+//  Created by 森篤史 on 2018/05/20.
+//  Copyright © 2018年 Cyder. All rights reserved.
+//
+
+protocol BaseCoordinator: class {
+    func start()
+}
